Fix Unity 6000.3.19f1 import + C# compilation errors

+++ Assets/Game/Core/Events/EventBus.cs
@@ -137,7 +137,7 @@ namespace Relicfall.Core.Events
     {
         public UnityEngine.Vector3 DeathPosition;
         public string DeathCause;
-        public int RunDurationSeconds;
+        public float RunDurationSeconds;
     }
 
     // Player Events
